FindAuthor is allowed to return null if no authors are found

+++ src/Chirp.Infrastructure/AuthorRepository.cs
@@ -45,27 +45,29 @@ public class AuthorRepository : IAuthorRepository
     public async Task<IEnumerable<AuthorDTO>> GetAllAuthors()
     {
 
-        return await GetAuthors(int.MaxValue,0);
+        return await GetAuthors(int.MaxValue, 0);
     }
 
     public async Task<AuthorDTO> FindAuthorByName(string author)
     {
 
+#pragma warning disable CS8603 // Possible null reference return.
         return await _context.Authors
         .Where(a => a.UserName!.Contains(author))
         .OrderByDescending(a => a.UserName!)
         .Select(a => new AuthorDTO(a!.UserName!, a.Email!, a.Cheeps.Select(c => new CheepDTO(c.Author.UserName!, c.Message, c.TimeStamp)))).SingleOrDefaultAsync();
+#pragma warning restore CS8603 // Possible null reference return.
     }
 
 
     public async Task<AuthorDTO> FindAuthorByEmail(string email)
     {
+#pragma warning disable CS8603 // Possible null reference return.
         return await _context.Authors
         .Where(a => a.Email != null && a.Email == email)
         .OrderByDescending(a => a.Email)
-        .Select(a => new AuthorDTO(a!.UserName, a.Email, a.Cheeps.Select(c => new CheepDTO(c.Author.UserName, c.Message, c.TimeStamp)))).SingleOrDefaultAsync();
-
-
+        .Select(a => new AuthorDTO(a!.UserName!, a.Email!, a.Cheeps.Select(c => new CheepDTO(c.Author.UserName!, c.Message, c.TimeStamp)))).SingleOrDefaultAsync();
+#pragma warning restore CS8603 // Possible null reference return.
     }
 
     public async Task DeleteAuthor(string author)
@@ -116,7 +118,7 @@ public class AuthorRepository : IAuthorRepository
     }
     public async Task<IEnumerable<AuthorDTO>> GetAllFollowed(string author)
     {
-        return await GetFollowed(author,int.MaxValue,0);
+        return await GetFollowed(author, int.MaxValue, 0);
     }
 
     //Finds who a specific author follows
@@ -134,11 +136,15 @@ public class AuthorRepository : IAuthorRepository
     }
     public async Task<IEnumerable<AuthorDTO>> GetAllFollowing(string author)
     {
-        return await GetFollowing(author,int.MaxValue,0);
+        return await GetFollowing(author, int.MaxValue, 0);
     }
 
     public async Task RemoveFollow(AuthorDTO followerDto, AuthorDTO followingDto)
     {
+        if (followerDto == null || followingDto == null)
+        {
+            return;
+        }
         var follow = await _context.Follows
         .SingleAsync(f => f.Following.UserName == followingDto.Name && f.Follower.UserName == followerDto.Name);
         _context.Follows.Remove(follow);
@@ -147,7 +153,10 @@ public class AuthorRepository : IAuthorRepository
 
     public async Task<bool> FollowExists(AuthorDTO followerDto, AuthorDTO followingDto)
     {
-
+        if (followerDto == null || followingDto == null)
+        {
+            return false;
+        }
         var follow = await _context.Follows
         .SingleOrDefaultAsync(f => f.Following.UserName == followingDto.Name && f.Follower.UserName == followerDto.Name);
